Create README.md
Hi guys, to those who might find this a little bit useful (reviewing for future IGCSE Exam), my is just one of the solution(which might work) for the pre-release material of paper 2/2022. In real exam, you just need to write the pseudocode version of this, however, it is always to step ahead and challenge yourself to code the pre-release material by using either python or java as in the actual exam, you will have a more general idea how this task works and question 1 will become a piece of cake:)). Anyways, GOOD LUCK and BEST WISHES if you are going through the IGCSE CS.